+++ src/main/java/com/nemonotfound/nemosmossyblocks/NemosMossyBlocks.java
@@ -18,9 +18,9 @@ public class NemosMossyBlocks implements ModInitializer {
 	public void onInitialize() {
 		log.info("Thank you for using Nemo's Mossy Blocks!");
 		ModBlocks.registerBlocks();
+		ModItems.registerItems();
 		ModItemGroups.registerItemGroups();
 		ModItemGroups.registerItemsToItemGroups();
-		ModItems.registerItems();
 		Entities.registerEntities();
 	}
 }