+++ Changelog.md
@@ -1,7 +1,9 @@
-# Changelog v1.4.3
+# Changelog v1.5 
 
-## Fixes
-- Mossy Fences do not connect to other fences
+## Additions
+- Added Mossy Chiseled Stone Bricks
+- Added Infested Mossy Chiseled Stone Bricks
+- Added Mossy Bricks, Brick Slab, Stairs and Wall
 
 ### Downloads
 [CurseForge](https://curseforge.com/minecraft/mc-mods/nemos-mossy-blocks) <br>